Virtualize the logger instantiation
Add a _logger function in Module::Logger, which creates the logger
instance, making it easier to subclass the Logger module.

Add a Logger::StdErr module, which dumps everything to STDERR and
supporst no min_level, i.e. it will print all levels. This module is
convenient to use during development with Kelp::Less.