Move config and routes initialization to abstract methods
It is sometimes needed to load extra configuration to the default
config.pl + environment. For example when running cli script you may
want to disable the logger and middleware, while preserving the
environment specific configuration.

This commit allows you to override the initialization of Config module.
It also provides and easy method 'process_mode' to load a configuration
file.

Consider:

    # lib/MyApp.pm
    sub _load_config {
        my $self = shift;
        $self->SUPER::_load_config();

        # Load local configuration, if found
        if ( $self->local ) {
            $self->loaded_modules->{Config}->process_mode( $self->local );
        }
    }